Extracted unit tests from tests2.cpp
