+++ tests/tests.cpp
@@ -1,14 +1,6 @@
 #include <sqlite_orm/sqlite_orm.h>
 #include <catch2/catch_all.hpp>
 
-#ifdef _DEBUG
-#pragma comment(lib, "Catch2d.Lib")
-#pragma comment(lib, "manual-link/Catch2Maind.Lib")
-#else
-#pragma comment(lib, "Catch2.Lib")
-#pragma comment(lib, "manual-link/Catch2Main.Lib")
-#endif
-
 #include <vector>  //  std::vector
 #include <string>  //  std::string
 #include <cstdio>  //  remove
